Diversas modificações

+++ src/main/java/br/com/gerenciadorDeOficina/repositories/ServicoRepository.java
@@ -3,5 +3,5 @@ package br.com.gerenciadorDeOficina.repositories;
 import br.com.gerenciadorDeOficina.models.Servico;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ServicoRepository extends JpaRepository<Servico, Integer> {
+public interface ServicoRepository extends JpaRepository<Servico, Long> {
 }
